tweak: requested changes, NO SHARK SPAWN COUNTER ADDED
