+++ Content.Server/Dragon/Components/DragonComponent.cs
@@ -1,4 +1,3 @@
-using Content.Shared.Damage;
 using Content.Shared.NPC.Prototypes;
 using Robust.Shared.Audio;
 using Robust.Shared.Prototypes;
