+++ Content.Server/Dragon/Components/DragonRiftComponent.cs
@@ -37,8 +37,4 @@ public sealed partial class DragonRiftComponent : SharedDragonRiftComponent
 
     [ViewVariables(VVAccess.ReadWrite), DataField("spawn", customTypeSerializer: typeof(PrototypeIdSerializer<EntityPrototype>))]
     public string SpawnPrototype = "MobCarpDragon";
-
-    // Sunrise-Start
-
-    // Sunrise-End
 }
